pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minima-llm"
version = "0.1.0"
description = "Minimal async LLM backend with caching and batch execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Laura Dietz, University of New Hampshire" }
]
keywords = ["llm", "async", "openai", "batch", "caching"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core has NO dependencies - stdlib only
dependencies = []

[project.optional-dependencies]
dspy = ["dspy-ai>=2.0"]
yaml = ["pyyaml>=6.0"]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]
dev = [
    "minima-llm[dspy,yaml,test]",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/trec-auto-judge/minima-llm"
Documentation = "https://github.com/trec-auto-judge/minima-llm#readme"
Repository = "https://github.com/trec-auto-judge/minima-llm"

[tool.hatch.build.targets.wheel]
packages = ["src/minima_llm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py310"
